bot/requirements.txt
redis == 4.6.0
aiohttp == 3.8.5
Pillow==9.2.0
psycopg2==2.9.3
py-cord==2.1.3
python-dotenv==0.20.0
PyYAML==6.0.1
urllib3==1.26.10
requests==2.31.0
misskey.py